flask/requirements.txt
flask
weather-api
